Fix runtime error with cake clean if the output files do not exist yet.

+++ Cakefile
@@ -6,8 +6,11 @@ task 'build', 'compile source files to javascript', (options) ->
   compile 'bin/buildr.coffee', true
 
 task 'clean', 'delete compiled output', (options) ->
-  fs.unlink 'lib/buildr.js'
-  fs.unlink 'bin/buildr.js'
+  fs.exists 'lib/buildr.js', (exists) ->
+    fs.unlink 'lib/buildr.js' if exists
+
+  fs.exists 'bin/buildr.js', (exists) ->
+    fs.unlink 'bin/buildr.js' if exists
 
 compile = (sourceFile, needShebang) ->
   fs.readFile sourceFile, 'utf8', (error, coffeeCode) ->
